requirements.txt
flask==2.3.3
anthropic==0.18.1
python-dotenv==1.0.0
gunicorn==21.2.0
httpx>=0.24.1
flask-cors==4.0.0 